+++ src/mailman/bin/mailman.py
@@ -24,6 +24,7 @@ from functools import cmp_to_key
 from mailman import public
 from mailman.core.i18n import _
 from mailman.core.initialize import initialize
+from mailman.database.transaction import transaction
 from mailman.interfaces.command import ICLISubCommand
 from mailman.utilities.modules import find_components
 from mailman.version import MAILMAN_VERSION_FULL
@@ -95,4 +96,5 @@ def main():
                    else os.path.abspath(os.path.expanduser(args.config)))
     initialize(config_path)
     # Perform the subcommand option.
-    args.func(args)
+    with transaction():
+        args.func(args)